# src/Resources/schema/quote.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "com.shopware.quote",
    "version": "2026-04-08",
...
    "x-capability": "com.shopware.quote"
  },
...
  "security": [
    {
      "bearerAuth": [],
      "ucpAgent": []
    }
  ],
  "paths": {
    "/ucp/quotes": {
      "post": {
        "operationId": "requestQuote",
        "summary": "Request a quote (RFQ). Creates and submits the request in one call; the returned quote is in state `open`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/idempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quote created and submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/InsufficientScope"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "501": {
            "$ref": "#/components/responses/Unsupported"
          }
        }
      },
      "get": {
        "operationId": "listQuotes",
        "summary": "The customer's quotes, newest first. Only quotes owned by the customer the credential resolves to are listed. Use this to rediscover a quote whose id was not stored, or to poll several negotiations at once.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 25
            },
            "description": "Page size; values above 50 are clamped."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of quotes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/InsufficientScope"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "501": {
            "$ref": "#/components/responses/Unsupported"
          }
        }
      }
    },
    "/ucp/quotes/{id}": {
      "get": {
        "operationId": "getQuote",
        "summary": "Read a quote snapshot (state, line items with offered prices, totals, expiration_date, comments). Only quotes owned by the resolved customer are visible; anything else is 404.",
        "parameters": [
          {
            "$ref": "#/components/parameters/quoteId"
          }
        ],
        "responses": {
          "200": {
            "description": "Quote snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/InsufficientScope"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "501": {
            "$ref": "#/components/responses/Unsupported"
          }
        }
      }
    },
    "/ucp/quotes/{id}/counter": {
      "post": {
        "operationId": "counterQuote",
        "summary": "Counter-offer: new per-line requested unit prices and/or a comment. Valid only in state `replied`; transitions to `change_requested`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/quoteId"
          },
          {
            "$ref": "#/components/parameters/idempotencyKey"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CounterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated quote (state change_requested)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/InsufficientScope"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "501": {
            "$ref": "#/components/responses/Unsupported"
          }
        }
      }
    },
    "/ucp/quotes/{id}/accept": {
      "post": {
        "operationId": "acceptQuote",
        "summary": "Accept the offer. Accepting IS ordering: executes the quote-to-order flow and returns the order reference. Valid only in state `replied` and before expiration_date.",
        "parameters": [
          {
            "$ref": "#/components/parameters/quoteId"
          },
          {
            "$ref": "#/components/parameters/idempotencyKey"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order placed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Not acceptable in current state or expired (code CHECKOUT__QUOTE_CANNOT_PLACE_ORDER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/InsufficientScope"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "501": {
            "$ref": "#/components/responses/Unsupported"
          }
        }
      }
    },
    "/ucp/quotes/{id}/decline": {
      "post": {
        "operationId": "declineQuote",
        "summary": "Decline the offer with an optional comment. Valid only in state `replied`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/quoteId"
          },
          {
            "$ref": "#/components/parameters/idempotencyKey"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Declined quote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/InsufficientScope"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "501": {
            "$ref": "#/components/responses/Unsupported"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OAuth 2.0",
        "description": "Identity-linking access token. Obtain it through the account-linking flow in the prose spec: a signed POST /ucp/quote-agent/authorization-requests, the customer's consent on the returned authorization_url, then the code exchange at the token_endpoint. Calling the authorization_endpoint directly from an agent fails with identity_required. Its subject identifies the customer, which is what this shop authorizes on; request only scopes listed in scopes_supported. Where com.shopware.quote:manage is listed, request it together with dev.ucp.shopping.order:manage: the quote operations require the first, and accepting requires both."
      },
      "ucpAgent": {
        "type": "apiKey",
        "in": "header",
        "name": "UCP-Agent",
        "description": "Standard UCP agent identification header (profile URL). Signature requirements follow the sales channel signature policy."
      }
    },
    "parameters": {
      "quoteId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-f]{32}$"
        },
        "description": "Quote id (UUID hex). Ids are unguessable but treated as secrets of the owning customer: foreign ids return 404."
      },
      "idempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Replay guard for a mutating request. Required by shops whose signature policy demands it; a replayed response carries `Idempotency-Replay: 1`."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Access token missing, invalid, expired, or revoked",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Invalid payload, per-line price asks sent to a shop that cannot record them, the access token does not resolve to a customer, or Quote Management is not enabled for the customer (`customer_specific_features` must contain `{\"QUOTE_MANAGEMENT\": true}`)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Quote unknown or owned by another customer, without confirming which",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unsupported": {
        "description": "The quote capability is disabled, not licensed, or the commercial Quote Management backend is unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency key reused for a different request, or for a response that is no longer replayable (code `idempotency_conflict`)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InsufficientScope": {
        "description": "insufficient_scope: the access token is valid but lacks a scope this operation requires. Only on shops that list com.shopware.quote:manage in scopes_supported. Link the account again, requesting the scopes the message names.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "QuoteRequest": {
        "type": "object",
        "required": [
          "line_items"
        ],
        "properties": {
          "line_items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "description": "product_id identifies the Shopware product to quote",
              "properties": {
                "product_id": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1
                },
                "requested_unit_price": {
                  "type": "number",
                  "description": "Buyer's asking price per unit, in the shop's customer-facing currency"
                }
              },
              "required": [
                "product_id",
                "quantity"
              ]
            }
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "CounterRequest": {
        "type": "object",
        "properties": {
          "line_items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Reference an existing quote line by id (preferred) or product_id",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Quote line item id from GET /ucp/quotes/{id}"
                },
                "product_id": {
                  "type": "string"
                },
                "requested_unit_price": {
                  "type": "number",
                  "description": "New asking price per unit"
                }
              },
              "required": [
                "requested_unit_price"
              ]
            },
            "description": "Omit entirely on a shop that cannot record per-line asks — sending any line item there is a 422, and the ask belongs in `comment` instead."
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "DeclineRequest": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          }
        }
      },
      "Quote": {
        "type": "object",
        "required": [
          "id",
          "state",
          "expiration_date",
          "totals",
          "line_items"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "quote_number": {
            "type": "string"
          },
          "state": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "draft",
              "open",
              "in_review",
              "replied",
              "change_requested",
              "accepted",
              "declined",
              "expired",
              "withdrawn",
              "cancelled",
              "reopen",
              null
            ],
            "description": "Null only if the merchant's state machine state could not be resolved. `draft` is a request left behind by a quote-to-order flow that never reached the merchant."
          },
          "expiration_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Always present. Accept before this instant; null means the merchant has not set one yet (the quote may still be expired by a background job)."
          },
          "currency": {
            "type": [
              "string",
              "null"
            ],
            "description": "ISO 4217 code; all amounts in this document are in this currency"
          },
          "totals": {
            "type": "object",
            "properties": {
              "gross": {
                "type": [
                  "number",
                  "null"
                ],
                "description": "Total including tax"
              },
              "net": {
                "type": [
                  "number",
                  "null"
                ],
                "description": "Total excluding tax"
              },
              "tax_status": {
                "type": [
                  "string",
                  "null"
                ],
                "enum": [
                  "gross",
                  "net",
                  "tax-free",
                  null
                ],
                "description": "Which total is the customer-facing authoritative amount; null if the commercial backend did not report one"
              }
            }
          },
          "line_items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "product_id": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "label": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "unit_price": {
                  "type": "number",
                  "description": "Current (offered) price per unit"
                },
                "total_price": {
                  "type": "number"
                },
                "requested_unit_price": {
                  "type": [
                    "number",
                    "null"
                  ],
                  "description": "Buyer's last ask per unit, if any. Always null on a shop that cannot record per-line asks (see Price semantics)."
                }
              }
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "comment": {
                  "type": "string"
                },
                "author": {
                  "type": "string",
                  "enum": [
                    "buyer",
                    "merchant"
                  ]
                },
                "created_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                }
              }
            }
          },
          "order": {
            "type": "object",
            "description": "Present only after acceptance",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "order_number": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          },
          "a2cn_session_id": {
            "type": "string",
            "description": "Present only when the quote was created with an A2CN session stamped onto it. The buyer's negotiation session id for POST /a2cn/sessions/{sessionId}/messages; derivable from the quote id, so its absence is not an error."
          },
          "ucp": {
            "type": "object",
            "description": "Protocol envelope added to every successful response; `capabilities` appears once an agent has negotiated.",
            "properties": {
              "version": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "const": "success"
              },
              "capabilities": {
                "type": "object"
              }
            }
          }
        }
      },
      "AcceptResponse": {
        "description": "The accepted quote, with the order block populated.",
        "$ref": "#/components/schemas/Quote"
      },
      "Error": {
        "type": "object",
        "required": [
          "ucp",
          "messages"
        ],
        "properties": {
          "ucp": {
            "type": "object",
            "description": "Protocol envelope. `status` is `error`; the failure itself is in `messages`.",
            "properties": {
              "version": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "const": "error"
              }
            }
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "code": {
                  "type": "string",
                  "description": "Machine-readable error code",
                  "examples": [
                    "validation_error",
                    "resource_not_found",
                    "unsupported_capability"
                  ]
                },
                "content": {
                  "type": "string"
                },
                "severity": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Standard UCP error envelope produced by the SDK (validation, not-found, idempotency-conflict and unsupported-capability errors). There is no top-level `status`: the status lives in `ucp`."
      },
      "QuoteList": {
        "type": "object",
        "required": [
          "quotes",
          "total",
          "limit",
          "page"
        ],
        "properties": {
          "quotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quote"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total quotes for this customer, not just this page"
          },
          "limit": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          }
        }
      }
    }
  }
}